- fixing unix build for points and poinths ... - plus some other stuff
git-svn-id: https://svn.code.sf.net/p/lass/code/lass/trunk@80 e9412239-ebb7-49aa-8f4a-f9aa784616d7

+++ src/num/distribution.h
@@ -80,7 +80,7 @@ private:
 	TGenerator* generator_;
 	TValue inf_;
 	TValue sup_;
-	double scale_;
+	long double scale_;
 };
